Create `Jekyll::Tags::IncludeTagBase` class
